Skip broken Firefox test
When exercising the polyfill in Firefox, the invalid JSON parsing error
will cripple the test suite, causing cascading failures in Mocha HTML
output as well as cause failures in unrelated tests.

The code fix was proposed #157 but we decided not to pull it since it's
a Firefox bug and Firefox now ships with native fetch anyway. Instead,
we avoid running that test as part of our polyfill test suite.