Add support for ArrayBufferView as POST body
Section 6.2 of the spec says that body can be BufferSource.
BufferSource is defined in https://heycam.github.io/webidl/#BufferSource:

> The BufferSource typedef is used to represent objects that are either
> themselves an ArrayBuffer or which provide a view on to an ArrayBuffer.

Furthermore, passing an ArrayBuffer directly to XMLHttpRequest is
considered deprecated and Safari throws a warning about it. This makes
it even more important to support ArrayBufferView in fetch.

The fallback for `ArrayBuffer.isView` is provided for IE 10, which
doesn't implement the method.